Further refactor and simplify `ChannelScrollContainer`
